+++ l1-contracts/src/interfaces/IUniFiAVSManager.sol
@@ -234,5 +234,5 @@ interface IUniFiAVSManager {
      * @param chainId The chain ID to check.
      * @return bool True if the validator is registered for the given chain ID, false otherwise.
      */
-    function isValidatorInChainId(bytes32 blsPubKeyHash, uint32 chainId) external view returns (bool);
+    function isValidatorInChainId(bytes32 blsPubKeyHash, uint256 chainId) external view returns (bool);
 }
